requirements/base.txt
# Core dependencies for AI Vehicle Monitoring System

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector database and similarity search
faiss-cpu==1.7.4
chromadb==0.4.18
numpy==1.24.3
scipy==1.11.4

# LangChain and LLM integration
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
tiktoken==0.5.2

# Data processing and ML
pandas==2.1.4
scikit-learn==1.3.2
sentence-transformers==2.2.2
transformers==4.36.2

# Database and storage
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Async and concurrency
asyncio-mqtt==0.16.1
aiofiles==23.2.1
httpx==0.25.2

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
